Added dbhelper file Modified login post method
